# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /paypal/webhook:
    post:
      operationId: receivePayPalWebhook
      tags: [payments]
      summary: Receive a verified PayPal webhook
      description: Raw JSON only. The server verifies PayPal transmission headers through PayPal's verify-webhook-signature endpoint.
      responses:
        "202": { description: Event accepted }
        "401": { description: Invalid webhook signature }
...
  /me/invoices:
    get:
      operationId: getMyInvoices
      tags: [member]
      summary: Get the authenticated member's own invoices
      responses:
        "200":
          description: Personal invoice history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
        "401":
          description: Authentication required
  /me/invoices/{id}/pdf:
    get:
      operationId: downloadMyInvoicePdf
      tags: [member]
      summary: Download the authenticated member's own invoice PDF
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invoice PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "401":
          description: Authentication required
        "403":
          description: This invoice does not belong to the authenticated member
        "404":
          description: Invoice not found
  /admin/invoices/{id}/pdf:
    get:
      operationId: downloadAdminInvoicePdf
      tags: [moderation]
      summary: Download any invoice PDF (moderator/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invoice PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "401":
          description: Moderator identity required
        "403":
          description: Moderator role required
        "404":
          description: Invoice not found
...
components:
...
  schemas:
...
    Invoice:
      type: object
      properties:
        id: { type: string }
        invoiceNumber: { type: string }
        paymentId: { type: string }
        listingId: { type: ["string", "null"] }
        listingName: { type: ["string", "null"] }
        listingWebsiteUrl: { type: ["string", "null"] }
        campaignNumber: { type: ["number", "null"] }
        type: { type: string, enum: [OWNER_BID, COMMUNITY_BID, PENALTY, SPONSORSHIP] }
        status: { type: string, enum: [PAID, REFUNDED, PARTIALLY_REFUNDED] }
        currency: { type: string }
        subtotal: { type: number }
        tax: { type: number }
        platformFee: { type: number }
        total: { type: number }
        refunded: { type: number }
        billingName: { type: ["string", "null"] }
        billingEmail: { type: ["string", "null"] }
        provider: { type: string }
        reason: { type: ["string", "null"] }
        pdfAvailable: { type: boolean }
        issuedAt: { type: string }
      required: [id, invoiceNumber, paymentId, listingId, listingName, listingWebsiteUrl, campaignNumber, type, status, currency, subtotal, tax, platformFee, total, refunded, billingName, billingEmail, provider, reason, pdfAvailable, issuedAt]
...